Split scenario details, allow edit title, deprecate author/contact fields
The author and contact fields in the scenario format were
never made visible to the player at all, and scenario designers
previously couldn't even edit their author name later.

I think it makes sense to move that information somewhere
that it will be visible and well-formatted. So instead of
"Who wrote 1/Who wrote 2" I'm proposing to call those Credits
and Tagline, and suggest to designers how to format them.